+++ src/Resource/Render.php
@@ -6,18 +6,12 @@ class Render implements Resource
 {
     private $content;
 
-    public function __construct(string $content)
+    private $snippetId;
+
+    public function __construct(int $snippetId, string $content)
     {
+        $this->snippetId = $snippetId;
         $this->content = $content;
-    }
-
-    public static function from(array $payload): Render
-    {
-        if (!\array_key_exists('content', $payload)) {
-            throw new \BadMethodCallException("Key 'content' does not exist");
-        }
-
-        return new Render($payload['content']);
     }
 
     public function __toString(): string
@@ -25,8 +19,26 @@ class Render implements Resource
         return $this->content;
     }
 
+    public static function from(array $payload): Render
+    {
+        if (!\array_key_exists('snippet_id', $payload)) {
+            throw new \BadMethodCallException("Key 'snippet_id' does not exist");
+        }
+
+        if (!\array_key_exists('content', $payload)) {
+            throw new \BadMethodCallException("Key 'content' does not exist");
+        }
+
+        return new Render($payload['snippet_id'], $payload['content']);
+    }
+
     public function getContent(): string
     {
         return $this->content;
+    }
+
+    public function getSnippetId(): int
+    {
+        return $this->snippetId;
     }
 }