+++ tests/ClientTest.php
@@ -4,7 +4,6 @@ namespace Jahuty;
 
 use donatj\MockWebServer\{MockWebServer, Response};
 use Jahuty\Action\Show;
-use Jahuty\Cache\Ttl;
 use Jahuty\Resource\Render;
 use Psr\SimpleCache\CacheInterface;
 
@@ -23,14 +22,14 @@ class ClientTest extends \PHPUnit\Framework\TestCase
         self::$server->stop();
     }
 
-    public function testConstructThrowsExceptionWhenCacheInvalid(): void
+    public function testConstructThrowsExceptionWhenCacheIsInvalid(): void
     {
-        $this->expectException(\InvalidARgumentException::class);
+        $this->expectException(\TypeError::class);
 
         (new Client('foo', ['cache' => 'foo']));
     }
 
-    public function testConstructThrowsExceptionWhenTtlInvalid(): void
+    public function testConstructThrowsExceptionWhenTtlIsInvalid(): void
     {
         $this->expectException(\InvalidARgumentException::class);
 
@@ -50,53 +49,6 @@ class ClientTest extends \PHPUnit\Framework\TestCase
         $this->assertInstanceOf(
             Service\Snippet::class,
             (new Client('foo'))->snippets
-        );
-    }
-
-    public function testFetchReturnsResourceWhenSuccess(): void
-    {
-        $id      = 1;
-        $content = 'foo';
-
-        $this->setupEndpointWithSuccess($id, $content);
-
-        // Mock the cache to store the returned value.
-        $cache = $this->createMock(CacheInterface::class);
-        $cache->expects($this->once())->method('set');
-
-        $client = new Client('1234abcd', [
-            'base_uri' => self::$server->getServerRoot(),
-            'cache'    => $cache
-        ]);
-
-        $action = new Show('render', $id);
-
-        $expected = new Render($content);
-        $actual   = $client->fetch($action, $this->createMock(Ttl::class));
-
-        $this->assertEquals($expected, $actual);
-    }
-
-    public function testFetchThrowsExceptionWhenProblem(): void
-    {
-        $id = 1;
-
-        $this->expectException(Exception\Error::class);
-
-        $this->setupEndpointWithProblem($id);
-
-        // Mock the cache NOT TO store the returned problem.
-        $cache = $this->createMock(CacheInterface::class);
-        $cache->expects($this->never())->method('set');
-
-        $client = new Client('1234abcd', [
-            'base_uri' => self::$server->getServerRoot(),
-            'cache'    => $cache
-        ]);
-
-        $client->fetch(
-            new Show('render', $id),
-            $this->createMock(Ttl::class)
         );
     }
 
@@ -128,15 +80,42 @@ class ClientTest extends \PHPUnit\Framework\TestCase
 
         $action = new Show('render', $id);
 
-        $expected = new Render($content);
+        $expected = new Render($id, $content);
         $actual   = $client->request($action);
 
         $this->assertEquals($expected, $actual);
     }
 
+    public function testSetBaseUri(): void
+    {
+        $client = new Client('foo');
+
+        $this->assertSame($client, $client->setBaseUri('https://google.com'));
+    }
+
+    public function testSetCache(): void
+    {
+        $client = new Client('foo');
+
+        $cache = $this->createMock(CacheInterface::class);
+
+        $this->assertSame($client, $client->setCache($cache));
+    }
+
+    public function testSetTtl(): void
+    {
+        $client = new Client('foo');
+
+        $this->assertSame($client, $client->setTtl(60));
+    }
+
     private function setupEndpointWithSuccess($id = 1, $content = 'foo'): void
     {
-        $response = new Response('{"content":"'. $content .'"}', [], 200);
+        $response = new Response(
+            '{"snippet_id":'. $id .', "content":"'. $content .'"}',
+            ['Content-Type' => 'application/json'],
+            200
+        );
 
         self::$server->setResponseOfPath("snippets/$id/render", $response);
     }