+++ tests/Resource/RenderTest.php
@@ -8,7 +8,7 @@ class RenderTest extends \PHPUnit\Framework\TestCase
 
     public function setUp(): void
     {
-        $this->payload = ['content' => 'foo'];
+        $this->payload = ['snippet_id' => 1, 'content' => 'foo'];
     }
 
     public function testFromThrowsExceptionIfContentsDoesNotExist(): void
@@ -20,9 +20,28 @@ class RenderTest extends \PHPUnit\Framework\TestCase
         Render::from($this->payload);
     }
 
-    public function testFrom(): void
+    public function testFromThrowsExceptionWhenSnippetIdIsMissing(): void
     {
-        $expected = new Render('foo');
+        $this->expectException(\BadMethodCallException::class);
+
+        unset($this->payload['snippet_id']);
+
+        Render::from($this->payload);
+    }
+
+    public function testFromReturnsRenderWhenPayloadIsValid(): void
+    {
+        $expected = new Render(1, 'foo');
+        $actual   = Render::from($this->payload);
+
+        $this->assertEquals($expected, $actual);
+    }
+
+    public function testFromReturnsRenderWhenExtraAttributesPresent(): void
+    {
+        $this->payload['foo'] = 'bar';
+
+        $expected = new Render(1, 'foo');
         $actual   = Render::from($this->payload);
 
         $this->assertEquals($expected, $actual);
@@ -30,18 +49,23 @@ class RenderTest extends \PHPUnit\Framework\TestCase
 
     public function testGetContent(): void
     {
-        $this->assertEquals('foo', (new Render('foo'))->getContent());
+        $this->assertEquals('foo', (new Render(1, 'foo'))->getContent());
+    }
+
+    public function testGetSnippetId(): void
+    {
+        $this->assertEquals(1, (new Render(1, 'foo'))->getSnippetId());
     }
 
     public function testToString(): void
     {
-        $this->assertEquals('foo', (string)new Render('foo'));
+        $this->assertEquals('foo', (string)new Render(1, 'foo'));
     }
 
     // Required for caching.
     public function testObjectSpportsLosslessSerializationAndDeserialization(): void
     {
-        $render = new Render('foo');
+        $render = new Render(1, 'foo');
 
         $this->assertEquals($render, unserialize(serialize($render)));
     }