+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 5.2.0 - 2020-03-045
+
+- Add support for rendering and caching a collection of snippets via the `snippets->allRenders()` method.
+- Remove the `Cache\Manager` and `Service\Factory` and move caching to the `Service\Snippet`. We got too fancy, and it got in the way.
+
 ## 5.1.1 - 2020-12-27
 
 - Move `Jahuty\Ttl\Ttl` to `Jahuty\Cache\Ttl`.